# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SmartCook API
  version: 1.0.0
  description: HTTP API exposed by the SmartCook Nextcloud application.
servers:
  - url: /apps/smartcook
security:
  - nextcloudSession: []
...
paths:
  /recipes:
    get:
      tags: [Recipes]
      summary: List recipes accessible to the current user
      parameters:
        - { in: query, name: search, schema: { type: string } }
        - { in: query, name: favorite, schema: { type: boolean } }
        - { in: query, name: status, schema: { type: string } }
        - { in: query, name: difficulty, schema: { type: string } }
        - { in: query, name: maxTime, schema: { type: integer, minimum: 0 } }
        - { in: query, name: maxCalories, schema: { type: integer, minimum: 0 } }
        - { in: query, name: tags, schema: { type: string } }
        - { in: query, name: categories, schema: { type: string } }
        - { in: query, name: tools, schema: { type: string } }
        - { in: query, name: ingredients, schema: { type: string } }
        - { in: query, name: excludeAllergens, schema: { type: string } }
        - { in: query, name: sort, schema: { type: string } }
        - { in: query, name: direction, schema: { enum: [asc, desc] } }
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 200 } }
      responses:
        '200':
          description: Recipe summaries
          content:
            application/json:
              schema:
                type: object
                required: [recipes]
                properties:
                  recipes: { type: array, items: { $ref: '#/components/schemas/Recipe' } }
    post:
      tags: [Recipes]
      summary: Create a recipe
      requestBody: { $ref: '#/components/requestBodies/RecipeBody' }
      responses:
        '201': { $ref: '#/components/responses/RecipeResponse' }
        '422': { $ref: '#/components/responses/ErrorResponse' }
  /recipes/{id}:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [Recipes]
      summary: Get a detailed recipe
      responses:
        '200': { $ref: '#/components/responses/RecipeResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
    put:
      tags: [Recipes]
      summary: Update a recipe
      requestBody: { $ref: '#/components/requestBodies/RecipeBody' }
      responses:
        '200': { $ref: '#/components/responses/RecipeResponse' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
        '422': { $ref: '#/components/responses/ErrorResponse' }
    delete:
      tags: [Recipes]
      summary: Delete an owned recipe
      responses:
        '200': { $ref: '#/components/responses/DeletedResponse' }
  /recipes/{id}/favorite:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [Recipes]
      summary: Set favorite state
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [favorite]
              properties: { favorite: { type: boolean } }
      responses:
        '200': { $ref: '#/components/responses/RecipeResponse' }
  /recipes/{id}/cooked:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [Recipes]
      summary: Increment the cooked counter
      responses:
        '200': { $ref: '#/components/responses/RecipeResponse' }
  /recipes/{id}/versions:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [Recipes]
      summary: List revision snapshots
      responses:
        '200':
          description: Revisions
          content:
            application/json:
              schema:
                type: object
                properties:
                  versions: { type: array, items: { $ref: '#/components/schemas/RecipeVersion' } }
  /recipes/{id}/restore/{revision}:
    parameters:
      - { $ref: '#/components/parameters/Id' }
      - { in: path, name: revision, required: true, schema: { type: integer, minimum: 1 } }
    post:
      tags: [Recipes]
      summary: Restore a prior revision as the newest revision
      responses:
        '200': { $ref: '#/components/responses/RecipeResponse' }
  /duplicates/check:
    post:
      tags: [Recipes]
      summary: Find likely duplicate recipes
      requestBody: { $ref: '#/components/requestBodies/RecipeBody' }
      responses:
        '200':
          description: Similarity-ranked matches
          content:
            application/json:
              schema:
                type: object
                properties:
                  matches: { type: array, items: { $ref: '#/components/schemas/DuplicateMatch' } }
  /recipes/{id}/merge:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [Recipes]
      summary: Merge another recipe or inline data into a recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                incomingRecipeId: { type: integer, minimum: 1 }
                recipe: { $ref: '#/components/schemas/RecipeInput' }
      responses:
        '200': { $ref: '#/components/responses/RecipeResponse' }
...
components:
  securitySchemes:
    nextcloudSession:
      type: apiKey
      in: cookie
      name: nc_session_id
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: integer, minimum: 1 }
  requestBodies:
    RecipeBody:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [recipe]
            properties: { recipe: { $ref: '#/components/schemas/RecipeInput' } }
...
  responses:
    RecipeResponse:
      description: Recipe envelope
      content:
        application/json:
          schema:
            type: object
            required: [recipe]
            properties:
              recipe: { $ref: '#/components/schemas/Recipe' }
              share: { type: object, additionalProperties: true }
    DeletedResponse:
      description: Deleted
      content:
        application/json:
          schema:
            type: object
            properties: { ok: { type: boolean } }
    ErrorResponse:
      description: Error
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error: { type: string }
              errors: { type: object, additionalProperties: { type: string } }
  schemas:
    NamedEntity:
      type: object
      required: [name]
      properties:
        id: { type: integer }
        name: { type: string }
        color: { type: [string, 'null'] }
    Ingredient:
      type: object
      required: [name]
      properties:
        id: { type: integer }
        name: { type: string }
        normalizedName: { type: [string, 'null'] }
        originalText: { type: [string, 'null'] }
        quantity: { type: [string, 'null'] }
        amount: { type: [number, 'null'] }
        unit: { type: [string, 'null'] }
        notes: { type: [string, 'null'] }
        optional: { type: boolean }
        group: { type: [string, 'null'] }
        category: { type: [string, 'null'] }
        allergens: { type: array, items: { type: string } }
        substitutes: { type: array, items: { type: string } }
        sortOrder: { type: integer }
    RecipeStep:
      type: object
      required: [text]
      properties:
        id: { type: integer }
        text: { type: string }
        timerSeconds: { type: [integer, 'null'] }
        temperature: { type: [number, 'null'] }
        temperatureUnit: { type: [string, 'null'] }
        notes: { type: [string, 'null'] }
        imagePath: { type: [string, 'null'] }
        sortOrder: { type: integer }
    Media:
      type: object
      properties:
        id: { type: integer }
        kind: { type: string }
        path: { type: string }
        mime: { type: [string, 'null'] }
        altText: { type: [string, 'null'] }
    RecipeInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1, maxLength: 255 }
        subtitle: { type: [string, 'null'] }
        description: { type: [string, 'null'] }
        language: { type: string, default: en }
        author: { type: [string, 'null'] }
        sourceName: { type: [string, 'null'] }
        sourceUrl: { type: [string, 'null'], format: uri }
        license: { type: [string, 'null'] }
        status: { enum: [draft, published, archived] }
        visibility: { enum: [private, shared, public] }
        favorite: { type: boolean }
        servings: { type: integer, minimum: 1 }
        yieldText: { type: [string, 'null'] }
        prepTime: { type: integer, minimum: 0 }
        restTime: { type: integer, minimum: 0 }
        cookTime: { type: integer, minimum: 0 }
        totalTime: { type: integer, minimum: 0 }
        difficulty: { type: [string, 'null'] }
        costCents: { type: [integer, 'null'], minimum: 0 }
        currency: { type: [string, 'null'] }
        cuisine: { type: [string, 'null'] }
        course: { type: [string, 'null'] }
        mealType: { type: [string, 'null'] }
        cookingMethod: { type: [string, 'null'] }
        season: { type: [string, 'null'] }
        origin: { type: [string, 'null'] }
        calories: { type: [integer, 'null'], minimum: 0 }
        nutrition: { type: object, additionalProperties: true }
        notes: { type: [string, 'null'] }
        imagePath: { type: [string, 'null'] }
        ingredients: { type: array, items: { $ref: '#/components/schemas/Ingredient' } }
        steps: { type: array, items: { $ref: '#/components/schemas/RecipeStep' } }
        tags: { type: array, items: { $ref: '#/components/schemas/NamedEntity' } }
        categories: { type: array, items: { $ref: '#/components/schemas/NamedEntity' } }
        tools: { type: array, items: { $ref: '#/components/schemas/NamedEntity' } }
        media: { type: array, items: { $ref: '#/components/schemas/Media' } }
    Recipe:
      allOf:
        - { $ref: '#/components/schemas/RecipeInput' }
        - type: object
          properties:
            id: { type: integer }
            uuid: { type: string, format: uuid }
            ownerId: { type: string }
            revision: { type: integer }
            cookCount: { type: integer }
            createdAt: { type: integer }
            updatedAt: { type: integer }
    RecipeVersion:
      type: object
      properties:
        revision: { type: integer }
        snapshot: { $ref: '#/components/schemas/RecipeInput' }
        createdAt: { type: integer }
    DuplicateMatch:
      type: object
      properties:
        score: { type: number, minimum: 0, maximum: 1 }
        recipe: { $ref: '#/components/schemas/Recipe' }
...